# docs/swagger.yaml
basePath: /api/v1.
definitions:
  APIError:
    properties:
      data:
        $ref: '#/definitions/errorStruct'
      ok:
        type: boolean
    type: object
...
  APIResponse-MessageResponse:
    properties:
      data:
        $ref: '#/definitions/MessageResponse'
      ok:
        type: boolean
    type: object
...
  APIResponse-User:
    properties:
      data:
        $ref: '#/definitions/User'
      ok:
        type: boolean
    type: object
...
  APIResponse-array_Submission:
    properties:
      data:
        items:
          $ref: '#/definitions/Submission'
        type: array
      ok:
        type: boolean
    type: object
...
  APIResponse-array_User:
    properties:
      data:
        items:
          $ref: '#/definitions/User'
        type: array
      ok:
        type: boolean
    type: object
...
  LanguageConfig:
    properties:
      fileExtension:
        type: string
      id:
        type: integer
      language:
        type: string
      version:
        type: string
    type: object
  MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  Submission:
    properties:
      checkedAt:
        type: string
      id:
        type: integer
      language:
        $ref: '#/definitions/LanguageConfig'
      languageId:
        type: integer
      order:
        type: integer
      result:
        $ref: '#/definitions/SubmissionResult'
      status:
        type: string
      submittedAt:
        type: string
      task:
        $ref: '#/definitions/Task'
      taskId:
        type: integer
      user:
        $ref: '#/definitions/User'
      userId:
        type: integer
    type: object
  SubmissionResult:
    properties:
      code:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      message:
        type: string
      submissionId:
        type: integer
      testResults:
        items:
          $ref: '#/definitions/TestResult'
        type: array
    type: object
  Task:
    properties:
      createdAt:
        type: string
      createdBy:
        type: integer
      id:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  TestResult:
    properties:
      id:
        type: integer
      inputOutputId:
        type: integer
      passed:
        type: boolean
      submissionResultId:
        type: integer
    type: object
  User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/types.UserRole'
      surname:
        type: string
      username:
        type: string
    type: object
  UserChangePassword:
    properties:
      newPassword:
        maxLength: 50
        minLength: 8
        type: string
      newPasswordConfirm:
        maxLength: 50
        minLength: 8
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - newPasswordConfirm
    type: object
  UserEdit:
    properties:
      email:
        type: string
      name:
        type: string
      role:
        $ref: '#/definitions/types.UserRole'
      surname:
        type: string
      username:
        type: string
    type: object
  UserIDs:
    properties:
      userIDs:
        items:
          type: integer
        type: array
    type: object
...
  errorStruct:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  types.UserRole:
    enum:
    - student
    - teacher
    - admin
    type: string
    x-enum-varnames:
    - UserRoleStudent
    - UserRoleTeacher
    - UserRoleAdmin
info:
  contact: {}
  title: Mini-Maxit API
  version: 1.0.0
paths:
...
  /group/{id}/users:
    delete:
      consumes:
      - application/json
      description: Delete users from a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: User IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UserIDs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Delete users from a group
      tags:
      - group
    get:
      description: Get users in a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-array_User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Get users in a group
      tags:
      - group
    post:
      consumes:
      - application/json
      description: Add users to a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: User IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UserIDs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Add users to a group
      tags:
      - group
...
  /submission/user/{id}:
    get:
      description: If the user is a student, it fails with 403 Forbidden.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Limit the number of returned submissions
        in: query
        name: limit
        type: integer
      - description: Offset the returned submissions
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-array_Submission'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Get all submissions for a user
      tags:
      - submission
...
  /task/{id}/assign/users:
    post:
      description: Assigns a task to users by task ID and user IDs
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: User IDs
        in: body
        name: userIDs
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Assign a task to users
      tags:
      - task
...
  /task/{id}/unassign/users:
    delete:
      description: Unassigns a task from users by task ID and user IDs
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: User IDs
        in: body
        name: userIDs
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Unassign a task from users
      tags:
      - task
...
  /user/:
    get:
      description: Get all users
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-User'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Get all users
      tags:
      - user
  /user/{id}:
    get:
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Get user by ID
      tags:
      - user
    patch:
      description: Edit user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User Edit Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UserEdit'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Edit user
      tags:
      - user
  /user/{id}/password:
    patch:
      description: Change user password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User Change Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UserChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Change user password
      tags:
      - user
  /user/email:
    get:
      description: Get user by email
      parameters:
      - description: User email
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/APIResponse-User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/APIError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/APIError'
      summary: Get user by email
      tags:
      - user
...
swagger: "2.0"